+++ worlds/shapez_2/generate/files/upgrades.py
@@ -7,8 +7,8 @@ from ...data.items import buildings, island_buildings, mechanics, misc
 tabs: tuple[tuple[tuple[str, str], tuple[dict[str, ItemData], ...]], ...] = (
     (("BuildingReward", "BuildingDefinitionGroupId"), (buildings.always, buildings.starting,
                                                        buildings.simple_processors, buildings.sandbox)),
-    (("IslandGroupReward", "GroupId"), (island_buildings.always, island_buildings.starting)),
-    (("MechanicReward", "MechanicId"), (mechanics.always, mechanics.starting,
+    (("IslandGroupReward", "GroupId"), (island_buildings.always, island_buildings.starting, island_buildings.miners)),
+    (("MechanicReward", "MechanicId"), (mechanics.always, mechanics.starting, mechanics.starting,
                                         misc.task_lines, misc.operator_lines)),
 )