+++ worlds/shapez_2/generate/items/island_buildings.py
@@ -6,13 +6,27 @@ if TYPE_CHECKING:
 
 
 def generate_default(world: "Shapez2World") -> Iterator[Shapez2Item]:
-    from ...data.items.island_buildings import always
+    from ...data.items.island_buildings import always, miners
 
     for name, data in always.items():
         yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)
 
+    if "Unlock extensions with miners" not in world.options.item_pool_modifiers:
+        for name in ("Shape Miner Extension", "Fluid Miner Extension", "Fluid Miner"):
+            data = miners[name]
+            yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)
+    else:
+        name = "Fluid Miner + Extension"
+        data = miners[name]
+        yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)
 
-def generate_starting() -> Iterator[str]:
+
+def generate_starting(world: "Shapez2World") -> Iterator[str]:
     from ...data.items.island_buildings import starting
 
     yield from starting
+
+    if "Unlock extensions with miners" in world.options.item_pool_modifiers:
+        yield "Shape Miner + Extension"
+    else:
+        yield "Shape Miner"