+++ worlds/shapez_2/docs/changelog.md
@@ -10,7 +10,7 @@ Versions are sorted in ascending order, i.e. the most recent changes are at the 
 - Fixed current research points permanently being hidden
 - Added surprise out-of-logic item as upgrade (this is what actually fixes the research points visibility)
 - Added goal tab to manual-like client (this also fixes not being able to send the goal flag to the server)
-- Fixed multiworld output zip file being rejected by webhost if they contain a shapez 2 output zip
+- Fixed multiworld output zip files being rejected by webhost if they contain a shapez 2 output zip
 - Fixed `Maximum processors per milestone` in the `shape_generation_adjustments` option not working
 - Prevented milestones from having the same final shape twice 
 - Overhauled the classification of a bunch of items